refactor(export): Add AlgoliaIterator class
